# platform/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/catalog/snapshots/{iid}": {
      "get": {
        "summary": "Catalog Snapshots",
        "operationId": "catalog_snapshots_catalog_snapshots__iid__get",
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          },
          {
            "name": "start_ns",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223371976854775807,
              "minimum": 0,
              "title": "Start Ns"
            }
          },
          {
            "name": "end_ns",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223371976854775807,
              "minimum": 0,
              "title": "End Ns"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Catalog Snapshots Catalog Snapshots  Iid  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/errors": {
      "get": {
        "summary": "Errors",
        "description": "Every failure this process carried on past (`observability.error_ledger`). `counts`/`last`\nare this process's own in-memory tallies, unchanged since Story 23.1; `services` is every\nother process's durable ledger (story 23.3), read back from `ERROR_LEDGER_DIR` -- so a\ncollector's failure is visible here too, not only in its own container's log. Empty means\nnone since start; the frontend's error bar polls this so a malfunction cannot go unseen.\nFor a one-shot job's ledger (`error_ledger.is_job_service`, e.g. the nightly's\n`archive.<step>_<venue>`), every run writes a `process_start`, so its `since_start` is the\nlatest run's errors only (one file per venue, so one venue's run never hides another's); the\nfull history stays in its file (`?since_ns=` reads further back).",
        "operationId": "errors_api_errors_get",
        "parameters": [
          {
            "name": "since_ns",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since Ns"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/archive/status": {
      "get": {
        "summary": "Get Archive Status",
        "description": "503 until the bus has cached its first message -- the scheduler may be down or still\nstarting, and an invented \"never ran\" status would be a lie -- and 503 again once the\nscheduler's 30 s heartbeat has been silent for the bus's stale bound, so a dead `archive`\nservice never shows as a frozen \"ok\" or \"running\". Read through `buses` (not an imported\nname) so tests can swap the bus.",
        "operationId": "get_archive_status_api_archive_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveStatusResponse"
                }
              }
            }
          },
          "503": {
            "description": "No `archive:status` message received yet, or none recently"
          }
        }
      }
    },
...
    "/api/candles/{instrument_id}": {
      "get": {
        "summary": "Get Candles",
        "operationId": "get_candles_api_candles__instrument_id__get",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument Id"
            }
          },
          {
            "name": "before_ns",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Before Ns"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 120,
              "title": "Limit"
            }
          },
          {
            "name": "bar_seconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 60,
              "title": "Bar Seconds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandlesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/snapshots/{instrument_id}": {
      "get": {
        "summary": "Get Snapshots",
        "operationId": "get_snapshots_api_snapshots__instrument_id__get",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument Id"
            }
          },
          {
            "name": "before_ns",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Before Ns"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 900,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotSeriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArchiveRun": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "day": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          },
          "days": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days"
          },
          "started": {
            "type": "string",
            "title": "Started"
          },
          "finished": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/ArchiveStep"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "kind",
          "day",
          "started",
          "steps"
        ],
        "title": "ArchiveRun",
        "description": "One maintenance run. `finished` is null while it is running."
      },
...
      "ArchiveStatusResponse": {
        "properties": {
          "next_run": {
            "type": "string",
            "title": "Next Run"
          },
          "next_intraday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Intraday"
          },
          "running": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArchiveRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArchiveRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_intraday": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArchiveRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "backup": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "enabled",
                  "disabled"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Backup"
          }
        },
        "type": "object",
        "required": [
          "next_run",
          "last_run"
        ],
        "title": "ArchiveStatusResponse",
        "description": "`backup` (Story 26.1b) says whether the scheduler's full runs end in the off-site backup; null\nwhen the status predates the key."
      },
      "ArchiveStep": {
        "properties": {
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "exit": {
            "type": "integer",
            "title": "Exit"
          },
          "duration_s": {
            "type": "number",
            "title": "Duration S"
          }
        },
        "type": "object",
        "required": [
          "venue",
          "name",
          "exit",
          "duration_s"
        ],
        "title": "ArchiveStep"
      },
...
      "CandlesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CandleItem"
            },
            "type": "array",
            "title": "Items"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "venue": {
            "type": "string",
            "title": "Venue"
          },
          "market": {
            "type": "string",
            "title": "Market"
          },
          "price_precision": {
            "type": "integer",
            "title": "Price Precision"
          },
          "size_precision": {
            "type": "integer",
            "title": "Size Precision"
          }
        },
        "type": "object",
        "required": [
          "items",
          "has_more",
          "venue",
          "market",
          "price_precision",
          "size_precision"
        ],
        "title": "CandlesResponse"
      },
...
      "ErrorsResponse": {
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "last": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Last"
          },
          "services": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ServiceErrorSummary"
            },
            "type": "object",
            "title": "Services"
          }
        },
        "type": "object",
        "required": [
          "counts",
          "last",
          "services"
        ],
        "title": "ErrorsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ServiceErrorSummary": {
        "properties": {
          "last_start_ns": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Start Ns"
          },
          "since_start": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Since Start"
          },
          "since": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since"
          }
        },
        "type": "object",
        "required": [
          "last_start_ns",
          "since_start",
          "since"
        ],
        "title": "ServiceErrorSummary"
      },
      "SnapshotSeriesPoint": {
        "properties": {
          "t": {
            "type": "integer",
            "title": "T"
          },
          "bid_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Units"
          },
          "ask_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ask Units"
          },
          "price_precision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Precision"
          },
          "mid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mid"
          },
          "micro": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Micro"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          }
        },
        "type": "object",
        "required": [
          "t"
        ],
        "title": "SnapshotSeriesPoint"
      },
      "SnapshotSeriesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SnapshotSeriesPoint"
            },
            "type": "array",
            "title": "Items"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "venue": {
            "type": "string",
            "title": "Venue"
          },
          "market": {
            "type": "string",
            "title": "Market"
          }
        },
        "type": "object",
        "required": [
          "items",
          "has_more",
          "venue",
          "market"
        ],
        "title": "SnapshotSeriesResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}